Pull template docs generation out of markdown into general docs generation
